requirements.txt
bleach==6.2.0
bokeh==3.6.2
certifi==2024.8.30
charset-normalizer==3.4.0
clickhouse-connect==0.8.9
colorama==0.4.6
colorcet==3.1.0
contourpy==1.3.1
holoviews==1.20.0
idna==3.10
Jinja2==3.1.4
linkify-it-py==2.0.3
lz4==4.3.3
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
panel==1.5.4
param==2.1.1
pillow==11.0.0
python-dateutil==2.9.0.post0
pytz==2024.2
pyviz_comms==3.0.3
PyYAML==6.0.2
requests==2.32.3
six==1.17.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
webencodings==0.5.1
xyzservices==2024.9.0
zstandard==0.23.0
